updated and refactored row button generating for courses, changed view name
