+++ public/src/pages/StudentPage.tsx
@@ -11,7 +11,7 @@ import { INavInfo } from "../NavigationHelper";
 import { View, ViewPage } from "./ViewPage";
 import { EnrollmentView } from "./views/EnrollmentView";
 import { showLoader } from '../loader';
-import { ArchiveView } from "./views/ArchiveView";
+import { CourseVisibilityView } from "./views/VisibilityView";
 
 export class StudentPage extends ViewPage {
     private navMan: NavigationManager;
@@ -47,7 +47,7 @@ export class StudentPage extends ViewPage {
         this.navHelper.registerFunction<any>("courses/{courseid:number}/members", this.members);
         this.navHelper.registerFunction<any>("courses/{courseid:number}/{page}", this.courseMissing);
         this.navHelper.registerFunction<any>("enroll", this.enroll);
-        // this.navHelper.registerFunction<any>("archive", this.archive);
+        this.navHelper.registerFunction<any>("archive", this.archive);
     }
 
     public checkAuthentication(): boolean {
@@ -96,12 +96,12 @@ export class StudentPage extends ViewPage {
         }
         return <div>
             <h1>Course visibility</h1>
-            <ArchiveView
+            <CourseVisibilityView
                 enrollments={await this.courseMan.getAllUserEnrollments(curUser.getId())}
                 onChangeClick={(enrol: Enrollment) => {
                     return this.courseMan.changeCourseVisibility(enrol);
                 }}
-            ></ArchiveView>
+            ></CourseVisibilityView>
         </div>
     }
